app/services/crawl_service.py
import httpx
from bs4 import BeautifulSoup
from typing import Dict, Any, Optional
import asyncio

class CrawlService:
    def __init__(self):
        self.base_url = "https://diemthi.tuyensinh247.com"
    
    async def crawl_score_by_sbd(self, sbd: str) -> Optional[Dict[str, Any]]:
        """Crawl điểm thi theo số báo danh"""
        try:
            async with httpx.AsyncClient() as client:
                # URL pattern cho tra cứu điểm (cần research thêm)
                url = f"{self.base_url}/tra-cuu-diem-thi"
                response = await client.get(url, params={"sbd": sbd})
                
                if response.status_code == 200:
                    # Parse HTML response (implement sau khi research cấu trúc)
                    soup = BeautifulSoup(response.text, 'html.parser')
                    # TODO: Parse actual data structure
                    return {
                        "sbd": sbd,
                        "found": True,
                        "data": "Sample score data - cần implement thực tế"
                    }
        except Exception as e:
            print(f"Error crawling score for SBD {sbd}: {e}")
        
        return {"sbd": sbd, "found": False, "data": None}
    
    async def crawl_school_info(self, school_name: str) -> Optional[Dict[str, Any]]:
        """Crawl thông tin trường"""
        try:
            # TODO: Implement school information crawling
            return {
                "school_name": school_name,
                "found": True,
                "data": "Sample school data - cần implement thực tế"
            }
        except Exception as e:
            print(f"Error crawling school info: {e}")
            return None

crawl_service = CrawlService()