requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
httpx==0.25.2
beautifulsoup4==4.12.2
openai==1.3.7